Refactor zoom state and fns into Viewer context and utils
